Move code from channel/distr/common/testing to /common/testing
