+++ pkg/channel/distributed/dispatcher/dispatcher/dispatcher_test.go
@@ -30,8 +30,8 @@ import (
 	consumertesting "knative.dev/eventing-kafka/pkg/channel/distributed/common/kafka/consumer/testing"
 	consumerwrapper "knative.dev/eventing-kafka/pkg/channel/distributed/common/kafka/consumer/wrapper"
 	"knative.dev/eventing-kafka/pkg/channel/distributed/common/metrics"
-	commontesting "knative.dev/eventing-kafka/pkg/channel/distributed/common/testing"
 	commonclient "knative.dev/eventing-kafka/pkg/common/client"
+	commontesting "knative.dev/eventing-kafka/pkg/common/testing"
 	eventingduck "knative.dev/eventing/pkg/apis/duck/v1"
 	"knative.dev/pkg/logging"
 	logtesting "knative.dev/pkg/logging/testing"
